[refactor] PostController에서 ApiResponse -> CustomApiResponse 이름 변경 및 GroupRepository 추가
