+++ src/main/java/com/hana4/ggumtle/repository/GroupRepository.java
@@ -1,0 +1,8 @@
+package com.hana4.ggumtle.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.hana4.ggumtle.model.entity.group.Group;
+
+public interface GroupRepository extends JpaRepository<Group, Long> {
+}
