+++ src/main/java/com/hana4/ggumtle/service/PostService.java
@@ -7,8 +7,8 @@ import com.hana4.ggumtle.dto.post.PostResponseDto;
 import com.hana4.ggumtle.global.error.CustomException;
 import com.hana4.ggumtle.global.error.ErrorCode;
 import com.hana4.ggumtle.model.entity.group.Group;
-import com.hana4.ggumtle.model.entity.group.GroupRepository;
 import com.hana4.ggumtle.model.entity.user.User;
+import com.hana4.ggumtle.repository.GroupRepository;
 import com.hana4.ggumtle.repository.PostRepository;
 
 import jakarta.transaction.Transactional;
