+++ src/test/java/com/hana4/ggumtle/model/entity/post/PostRepositoryTest.java
@@ -13,8 +13,8 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import com.hana4.ggumtle.dto.user.UserRequestDto;
 import com.hana4.ggumtle.model.entity.group.Group;
 import com.hana4.ggumtle.model.entity.group.GroupCategory;
-import com.hana4.ggumtle.model.entity.group.GroupRepository;
 import com.hana4.ggumtle.model.entity.user.User;
+import com.hana4.ggumtle.repository.GroupRepository;
 import com.hana4.ggumtle.repository.PostRepository;
 import com.hana4.ggumtle.repository.UserRepository;
